Turns out that on Linux, more than VULKAN_FRAME_COUNT swap chain images are created, and now that's handled properly

+++ platform/unix/platform.c
@@ -57,7 +57,7 @@ Return Value:
 }
 
 PCSTR
-PlatCaptureStackBackTrace(_In_ UINT64 FramesToSkip, _In_ UINT64 MaxFrames)
+PlatCaptureStackBackTrace(_In_ UINT32 FramesToSkip, _In_ UINT32 MaxFrames)
 /*++
 
 Routine Description:
